requirements.txt
# Core ML Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Deep Learning (optional)
torch==2.1.2
torchvision==0.16.2

# MLOps & Experiment Tracking
mlflow==2.9.2
feast==0.34.0
great-expectations==0.18.8
evidently==0.4.11

# API & Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.2
redis==5.0.1

# Workflow Orchestration
apache-airflow==2.8.1
celery==5.3.4

# Data Processing
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyarrow==14.0.2

# Monitoring & Observability
prometheus-client==0.19.0
grafana-api==1.0.3

# Hyperparameter Tuning
optuna==3.5.0
hyperopt==0.2.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
jupyter==1.0.0
notebook==7.0.6
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
pyyaml==6.0.1
requests==2.31.0
kaggle==1.5.16

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Docker & Container Support
docker>=4.0.0,<7.0.0